# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SkillCoach AI · Backend",
    "description": "條件式 RAG、客戶對練、風險初篩與背景動態評分。COACH_AI=local 啟用本機模型。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/personas": {
      "get": {
        "tags": [
          "training"
        ],
        "summary": "Personas",
        "operationId": "personas_personas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Personas Personas Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/documents": {
      "get": {
        "tags": [
          "knowledge"
        ],
        "summary": "Documents",
        "operationId": "documents_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentInfo"
                  },
                  "type": "array",
                  "title": "Response Documents Documents Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Upload",
        "operationId": "upload_documents_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_documents_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/documents/demo": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Load Demo",
        "operationId": "load_demo_documents_demo_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/search": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Search",
        "operationId": "search_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ask": {
      "post": {
        "tags": [
          "knowledge"
        ],
        "summary": "Ask",
        "operationId": "ask_ask_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Chat",
        "operationId": "chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnChatResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/evaluate": {
      "post": {
        "tags": [
          "training"
        ],
        "summary": "Evaluate",
        "operationId": "evaluate_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions": {
      "post": {
        "tags": [
          "training loop"
        ],
        "summary": "Start Session",
        "operationId": "start_session_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/turns": {
      "post": {
        "tags": [
          "training loop"
        ],
        "summary": "Turn",
        "operationId": "turn_sessions__session_id__turns_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TurnRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}": {
      "get": {
        "tags": [
          "training loop"
        ],
        "summary": "Session State",
        "operationId": "session_state_sessions__session_id__get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/finish": {
      "post": {
        "tags": [
          "training loop"
        ],
        "summary": "Finish Session",
        "operationId": "finish_session_sessions__session_id__finish_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Gateway"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AskRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AskRequest"
      },
      "AskResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          },
          "is_mock": {
            "type": "boolean",
            "title": "Is Mock"
          },
          "insufficient_evidence": {
            "type": "boolean",
            "title": "Insufficient Evidence"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "sources",
          "is_mock",
          "insufficient_evidence"
        ],
        "title": "AskResponse"
      },
      "Body_upload_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_documents_post"
      },
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message"
          },
          "persona": {
            "type": "string",
            "enum": [
              "cautious",
              "fee_sensitive",
              "short_term"
            ],
            "title": "Persona",
            "default": "cautious"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "maxItems": 24,
            "title": "History"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
...
      "DocumentInfo": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "chunks": {
            "type": "integer",
            "title": "Chunks"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "chunks"
        ],
        "title": "DocumentInfo"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "EvaluateRequest": {
        "properties": {
          "history": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "maxItems": 24,
            "minItems": 1,
            "title": "History"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "history"
        ],
        "title": "EvaluateRequest"
      },
      "EvaluateResponse": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "scores": {
            "items": {
              "$ref": "#/components/schemas/DimensionScore"
            },
            "type": "array",
            "title": "Scores"
          },
          "improvements": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Improvements"
          },
          "suggested_reply": {
            "type": "string",
            "title": "Suggested Reply"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          },
          "is_mock": {
            "type": "boolean",
            "title": "Is Mock"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "scores",
          "improvements",
          "suggested_reply",
          "sources",
          "is_mock"
        ],
        "title": "EvaluateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestResponse": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "chunks": {
            "type": "integer",
            "title": "Chunks"
          },
          "duplicate": {
            "type": "boolean",
            "title": "Duplicate"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "chunks",
          "duplicate"
        ],
        "title": "IngestResponse"
      },
      "Message": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Content"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "Message"
      },
      "SearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Query"
          },
          "top_k": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchRequest"
      },
      "SearchResponse": {
        "properties": {
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "sources"
        ],
        "title": "SearchResponse"
      },
      "SessionRequest": {
        "properties": {
          "persona": {
            "type": "string",
            "enum": [
              "cautious",
              "fee_sensitive",
              "short_term"
            ],
            "title": "Persona",
            "default": "cautious"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SessionRequest"
      },
      "SessionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "persona": {
            "type": "string",
            "title": "Persona"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "finishing",
              "finished",
              "final_failed"
            ],
            "title": "Status"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/Message"
            },
            "type": "array",
            "title": "History"
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/StoredTurn"
            },
            "type": "array",
            "title": "Turns"
          },
          "compliance": {
            "items": {
              "$ref": "#/components/schemas/ComplianceFlag"
            },
            "type": "array",
            "title": "Compliance"
          },
          "evaluations": {
            "items": {
              "$ref": "#/components/schemas/EvaluationSnapshot"
            },
            "type": "array",
            "title": "Evaluations"
          },
          "latest_evaluation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluationSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "final_report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EvaluateResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "final_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "persona",
          "status",
          "history",
          "turns",
          "compliance",
          "evaluations",
          "latest_evaluation",
          "final_report",
          "final_error"
        ],
        "title": "SessionResponse"
      },
...
      "TurnChatResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          },
          "is_mock": {
            "type": "boolean",
            "title": "Is Mock"
          },
          "rag_used": {
            "type": "boolean",
            "title": "Rag Used"
          },
          "evidence_status": {
            "type": "string",
            "enum": [
              "not_needed",
              "retrieved_context",
              "missing"
            ],
            "title": "Evidence Status"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "sources",
          "is_mock",
          "rag_used",
          "evidence_status"
        ],
        "title": "TurnChatResponse"
      },
      "TurnRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "TurnRequest"
      },
      "TurnResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "type": "array",
            "title": "Sources"
          },
          "is_mock": {
            "type": "boolean",
            "title": "Is Mock"
          },
          "rag_used": {
            "type": "boolean",
            "title": "Rag Used"
          },
          "evidence_status": {
            "type": "string",
            "enum": [
              "not_needed",
              "retrieved_context",
              "missing"
            ],
            "title": "Evidence Status"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "turn": {
            "type": "integer",
            "title": "Turn"
          },
          "compliance": {
            "items": {
              "$ref": "#/components/schemas/ComplianceFlag"
            },
            "type": "array",
            "title": "Compliance"
          },
          "evaluation_status": {
            "type": "string",
            "const": "pending",
            "title": "Evaluation Status"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "sources",
          "is_mock",
          "rag_used",
          "evidence_status",
          "session_id",
          "turn",
          "compliance",
          "evaluation_status"
        ],
        "title": "TurnResponse"
      },
...
    }
  }
}